update makefile to support Java 8 lambdas
Java 8 lambdas use invokedynamic, which Avian has supported for some
time, but I never updated hello-ios's makefile to pass the `-hostvm`
option to the `bootimage-generator`, which is necessary to actually
AOT-compile bytecode containing invokedynamic.

This commit also cleans up the SDK-finding logic and selects Java 8 as
the default source and target languages.